fix: restore lin page, move modal preview to kakani page, remove outdated kakani tests
