+++ src/app/team/lin/page.tsx
@@ -1,57 +1,27 @@
 "use client";
 
 import { useState } from "react";
-import { AddMembersModal } from "@/components/groups/add-members-modal";
-import { Button } from "@/components/ui/button";
-
-const mockMembers = [
-  { memberId: 1, ownername: "Alice Johnson", isOwner: true, isSelected: true },
-  { memberId: 2, ownername: "Bob Smith", isOwner: false, isSelected: true },
-  { memberId: 3, ownername: "Carol White", isOwner: false, isSelected: false },
-  { memberId: 4, ownername: "David Brown", isOwner: false, isSelected: true },
-  { memberId: 5, ownername: "Eve Davis", isOwner: false, isSelected: false },
-  { memberId: 6, ownername: "Frank Miller", isOwner: false, isSelected: false },
-  { memberId: 7, ownername: "Grace Wilson", isOwner: false, isSelected: true },
-  { memberId: 8, ownername: "Hank Taylor", isOwner: false, isSelected: false },
-];
 
 export default function LinPage() {
-  const [open, setOpen] = useState(false);
-  const [submittingOpen, setSubmittingOpen] = useState(false);
-  const [members, setMembers] = useState(mockMembers);
+  const [count, setCount] = useState(0);
 
-  const handleSelectionChange = (memberId: number, selected: boolean) => {
-    setMembers((prev) => prev.map((m) => (m.memberId === memberId ? { ...m, isSelected: selected } : m)));
+  const increment = () => {
+    setCount(count + 1);
   };
 
   return (
-    <div className="p-20 flex flex-col items-center gap-4">
-      <h1 className="text-2xl font-bold mb-4">Add Members Modal Preview</h1>
-
-      <Button onClick={() => setOpen(true)}>Default State</Button>
-      <Button onClick={() => setSubmittingOpen(true)}>Submitting State</Button>
-
-      <AddMembersModal
-        open={open}
-        onOpenChange={setOpen}
-        groupName="Garden Club"
-        members={members}
-        onSelectionChange={handleSelectionChange}
-        onConfirm={() => {
-          alert("Saved!");
-          setOpen(false);
-        }}
-      />
-
-      <AddMembersModal
-        open={submittingOpen}
-        onOpenChange={setSubmittingOpen}
-        groupName="Garden Club"
-        members={members}
-        onSelectionChange={handleSelectionChange}
-        onConfirm={() => {}}
-        isSubmitting
-      />
+    <div className="flex flex-col items-center pt-12 gap-3">
+      <div className="font-bold text-[40px]">Kyle Lin</div>
+      <div className="font-bold text-[25px]">Developer</div>
+      <div>
+        <strong>Fun Fact: </strong>I like to dance
+      </div>
+      <button onClick={increment} className="bg-emerald-300 p-5 rounded-xl">
+        Click Me!{" "}
+      </button>
+      <div>
+        Count: <div data-testid="count">{count}</div>
+      </div>
     </div>
   );
 }
